Added support for Snort keyword 'fast_pattern'
Added support for the Snort 'fast_pattern' keyword which is used to force a
particular payload match to be done first.  This allows the signature author
to optimize the performance of certain signatures based on a knowledge of
how likely certain strings are to match within application layer protocols.
A gooo write up of the 'fast_pattern' keyword was posted to the VRT blog
here:

http://vrt-blog.snort.org/2010/04/using-snort-fast-patterns-wisely-for.html