Added content match ordering based on length
In cases where the 'fast_pattern' option is not used, Snort generally tries to
pick the longest pattern to match first since this should usually result in
better performance.  That is, longer there is a higher chance for a longer
string to be more unique, and this would result in shorter strings from not
being searched for.  This works in the context of iptables because 'matches'
are AND'd togther, so if the first string match fails, no subsequent string
matches will be executed.  Hence, the search for "shortstr" below would not
happen if the search for "thisisalongstring" failed:

-m string --string 'thisisalongstring' --algo bm -m string --string 'shortstr' --algo bm

One thing to note is that iptables does not support relative string matches
in the same way that Snort does.  The iptables string match can specify an
offset and depth into the packet via --from and --to.  The end result is that
the fwsnort way of maximizing performance is to find the longest string, do
the match, and apply an approximation for --from and --to whenever they are
required for any pattern.  That is, it doesn't have to worry about relative
matches and finding the end of a pattern in order to know where to start the
next search.  Now, this will result in signature matching in fwsnort not
being as accurate as Snort (remember that fwsnort emulates Snort behavior as
closely as possible given functionality implemented in iptables), but it
should be faster.